requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
astroid==4.0.2
async-generator==1.10
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
colorama==0.4.6
dill==0.4.1
distlib==0.4.0
filelock==3.20.3
h11==0.16.0
identify==2.6.16
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
isort==7.0.0
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.10.0
outcome==1.3.0.post0
packaging==26.0
platformdirs==4.5.1
pluggy==1.6.0
pre_commit==4.5.1
pycparser==3.0
Pygments==2.19.2
pylint==4.0.4
pylint-pytest==1.1.8
PySocks==1.7.1
pytest==8.2.0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
selenium==4.40.0
sniffio==1.3.1
sortedcontainers==2.4.0
timer==0.3.0
tomlkit==0.14.0
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing_extensions==4.15.0
urllib3==2.6.3
virtualenv==20.36.1
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.2
zipp==3.23.0
